+++ packages/datetime/src/getDatetimeRuleProcessorSQL.ts
@@ -14,6 +14,7 @@ import {
   isRelativeDateTimeValue,
   materializeRelativeValues,
   processIsDateField,
+  resolveDatetimeOperator,
 } from './utils';
 
 /**
@@ -117,7 +118,7 @@ export const getDatetimeRuleProcessorSQL =
   (apiFns: RQBDateTimeLibraryAPI): RuleProcessor =>
   (rule, options) => {
     const opts = options ?? /* v8 ignore start -- @preserve */ {} /* v8 ignore stop -- @preserve */;
-    const operator = mapSQLOperator(rule.operator);
+    const operator = mapSQLOperator(resolveDatetimeOperator(rule, opts));
     const operatorLowerCase = lc(operator);
     // v8 ignore next
     const {